Finished permissions for users

+++ routes/posts.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const Post = require('../models/post');
 
 // Middleware
-import { checkAuthenticated } from '../basicAuth';
+import { checkAuthenticated, authUser, renderEditDeleteButtons } from '../basicAuth';
 
 // Routes
 router.get('/new', checkAuthenticated, (req, res) => {
@@ -11,20 +11,26 @@ router.get('/new', checkAuthenticated, (req, res) => {
 	res.render('posts/new', { post: new Post(), user: req.user });
 });
 
-router.get('/edit/:id', async (req, res) => {
+router.get('/edit/:id', checkAuthenticated, async (req, res) => {
 	const post = await Post.findById(req.params.id);
-	res.status(200);
-	res.render('posts/edit', { post: post });
+	if (authUser(req.user, post)) {
+		res.status(200);
+		res.render('posts/edit', { post: post });
+	} else {
+		res.status(403);
+		res.render('invalid-permission');
+	}
 });
 
 router.get('/:slug', async (req, res) => {
-	const post = await Post.findOne({ slug: req.params.slug });
+	const post = await Post.findOne({ slug: req.params.slug }).populate('author');
 	if (post == null) {
 		res.status(404);
 		res.render('404', { error: null });
 	} else {
+		let renderButtons = renderEditDeleteButtons(req.user, post);
 		res.status(200);
-		res.render('posts/show', { post: post });
+		res.render('posts/show', { post: post, renderButtons: renderButtons });
 	}
 });
 
@@ -49,26 +55,36 @@ router.post('/', checkAuthenticated, async (req, res) => {
 	}
 });
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', checkAuthenticated, async (req, res) => {
 	let post = await Post.findById(req.params.id);
-	post.title = req.body.title;
-	post.description = req.body.description;
-	post.markdown = req.body.markdown;
-
-	try {
-		post = await post.save();
-		res.status(204);
-		res.redirect(`/posts/${post.slug}`);
-	} catch (error) {
-		res.status(500);
-		res.render('posts/edit', { post: post });
+	if (authUser(req.user, post)) {
+		post.title = req.body.title;
+		post.description = req.body.description;
+		post.markdown = req.body.markdown;
+		try {
+			post = await post.save();
+			res.status(204);
+			res.redirect(`/posts/${post.slug}`);
+		} catch (error) {
+			res.status(500);
+			res.render('posts/edit', { post: post });
+		}
+	} else {
+		res.status(403);
+		res.render('invalid-permission');
 	}
 });
 
-router.delete('/:id', async (req, res) => {
-	await Post.findByIdAndDelete(req.params.id);
-	res.status(204);
-	res.redirect('/');
+router.delete('/:id', checkAuthenticated, async (req, res) => {
+	const post = await Post.findById(req.params.id);
+	if (authUser(req.user, post)) {
+		await Post.findByIdAndDelete(req.params.id);
+		res.status(204);
+		res.redirect('/');
+	} else {
+		res.status(403);
+		res.render('invalid-permission');
+	}
 });
 
 module.exports = router;
